workwechat/requirements.txt
airtest==1.2.4
altgraph==0.17.2
cached-property==1.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
colorama==0.4.4
comtypes==1.1.10
configparser==5.2.0
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.4
docker==3.7.3
docker-compose==1.23.2
docker-pycreds==0.4.0
dockerpty==0.4.1
docopt==0.6.2
facebook-wda==1.4.6
filelock==3.4.2
future==0.18.2
gitdb==4.0.9
GitPython==3.1.26
greenlet==1.1.2
hrpc==1.0.9
idna==2.7
Jinja2==3.0.3
jsonschema==2.6.0
MarkupSafe==2.0.1
mss==6.1.0
Naked==0.1.31
nose==1.3.7
numpy==1.19.3
opencv-contrib-python==4.5.5.62
pefile==2021.9.3
Pillow==9.0.0
pipenv==2022.1.8
platformdirs==2.4.1
poco==0.97.0
pocoui==1.0.85
psutil==5.9.0
py==1.11.0
pyaml==21.10.1
pycparser==2.21
pycryptodome==3.12.0
PyGithub==1.55
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.5
PyJWT==2.3.0
PyNaCl==1.5.0
pypiwin32==223
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
python-dateutil==2.8.2
python-gitlab==1.15.0
pywin32==303
pywin32-ctypes==0.2.0
pywinauto==0.6.3
PyYAML==3.13
requests==2.20.1
retry==0.9.2
shellescape==3.8.1
six==1.16.0
smmap==5.0.0
SQLAlchemy==1.4.29
svn==1.0.1
texttable==0.9.1
urllib3==1.24.3
virtualenv==20.13.0
virtualenv-clone==0.5.7
websocket-client==0.48.0
wrapt==1.13.3
